PIM-3318: clean ORM filters and fix Mongo specs
