+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Filter/EntityFilter.php
@@ -2,7 +2,6 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter;
 
-use Doctrine\ORM\QueryBuilder;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\Operators;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\FieldFilterInterface;
 
@@ -13,16 +12,10 @@ use Pim\Bundle\CatalogBundle\Doctrine\Query\FieldFilterInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class EntityFilter implements FieldFilterInterface
+class EntityFilter extends AbstractFilter implements FieldFilterInterface
 {
-    /** @var QueryBuilder */
-    protected $qb;
-
     /** @var array */
     protected $supportedFields;
-
-    /** @var array */
-    protected $supportedOperators;
 
     /**
      * Instanciate the base filter
@@ -36,18 +29,6 @@ class EntityFilter implements FieldFilterInterface
     ) {
         $this->supportedFields    = $supportedFields;
         $this->supportedOperators = $supportedOperators;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setQueryBuilder($queryBuilder)
-    {
-        if (!($queryBuilder instanceof QueryBuilder)) {
-            throw new \InvalidArgumentException('Query builder should be an instance of Doctrine\ORM\QueryBuilder');
-        }
-
-        $this->qb = $queryBuilder;
     }
 
     /**
@@ -94,21 +75,5 @@ class EntityFilter implements FieldFilterInterface
     public function supportsField($field)
     {
         return in_array($field, $this->supportedFields);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
     }
 }
