+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Filter/PriceFilter.php
@@ -2,12 +2,8 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter;
 
-use Doctrine\ORM\QueryBuilder;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\Operators;
-use Pim\Bundle\CatalogBundle\Doctrine\ORM\Condition\CriteriaCondition;
-use Pim\Bundle\CatalogBundle\Doctrine\ORM\Join\ValueJoin;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\AttributeFilterInterface;
-use Pim\Bundle\CatalogBundle\Exception\ProductQueryException;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 
 /**
@@ -17,16 +13,10 @@ use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class PriceFilter implements AttributeFilterInterface
+class PriceFilter extends AbstractFilter implements AttributeFilterInterface
 {
-    /** @var QueryBuilder */
-    protected $qb;
-
     /** @var array */
     protected $supportedAttributes;
-
-    /** @var array */
-    protected $supportedOperators;
 
     /**
      * Instanciate the base filter
@@ -40,18 +30,6 @@ class PriceFilter implements AttributeFilterInterface
     ) {
         $this->supportedAttributes = $supportedAttributes;
         $this->supportedOperators  = $supportedOperators;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setQueryBuilder($queryBuilder)
-    {
-        if (!(($queryBuilder instanceof QueryBuilder))) {
-            throw new \InvalidArgumentException('Query builder should be an instance of Doctrine\ORM\QueryBuilder');
-        }
-
-        $this->qb = $queryBuilder;
     }
 
     /**
@@ -106,58 +84,6 @@ class PriceFilter implements AttributeFilterInterface
     public function supportsAttribute(AttributeInterface $attribute)
     {
         return in_array($attribute->getAttributeType(), $this->supportedAttributes);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
-    }
-
-    /**
-     * Prepare criteria condition with field, operator and value
-     *
-     * @param string|array $field    the backend field name
-     * @param string|array $operator the operator used to filter
-     * @param string|array $value    the value(s) to filter
-     *
-     * @return string
-     * @throws ProductQueryException
-     */
-    protected function prepareCriteriaCondition($field, $operator, $value)
-    {
-        $criteriaCondition = new CriteriaCondition($this->qb);
-
-        return $criteriaCondition->prepareCriteriaCondition($field, $operator, $value);
-    }
-
-    /**
-     * Prepare join to attribute condition with current locale and scope criterias
-     *
-     * @param AttributeInterface $attribute the attribute
-     * @param string             $joinAlias the value join alias
-     * @param string             $locale    the locale
-     * @param string             $scope     the scope
-     *
-     * @throws ProductQueryException
-     *
-     * @return string
-     */
-    protected function prepareAttributeJoinCondition(AttributeInterface $attribute, $joinAlias, $locale = null, $scope = null)
-    {
-        $joinHelper = new ValueJoin($this->qb);
-
-        return $joinHelper->prepareCondition($attribute, $joinAlias, $locale, $scope);
     }
 
     /**
