+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Filter/ProductIdFilter.php
@@ -2,10 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter;
 
-use Doctrine\ORM\QueryBuilder;
-use Pim\Bundle\CatalogBundle\Doctrine\ORM\Condition\CriteriaCondition;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\FieldFilterInterface;
-use Pim\Bundle\CatalogBundle\Exception\ProductQueryException;
 
 /**
  * Product id filter
@@ -14,16 +11,10 @@ use Pim\Bundle\CatalogBundle\Exception\ProductQueryException;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductIdFilter implements FieldFilterInterface
+class ProductIdFilter extends AbstractFilter implements FieldFilterInterface
 {
-    /** @var QueryBuilder */
-    protected $qb;
-
     /** @var array */
     protected $supportedFields;
-
-    /** @var array */
-    protected $supportedOperators;
 
     /**
      * Instanciate the base filter
@@ -37,18 +28,6 @@ class ProductIdFilter implements FieldFilterInterface
     ) {
         $this->supportedFields    = $supportedFields;
         $this->supportedOperators = $supportedOperators;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setQueryBuilder($queryBuilder)
-    {
-        if (!(($queryBuilder instanceof QueryBuilder))) {
-            throw new \InvalidArgumentException('Query builder should be an instance of Doctrine\ORM\QueryBuilder');
-        }
-
-        $this->qb = $queryBuilder;
     }
 
     /**
@@ -69,38 +48,5 @@ class ProductIdFilter implements FieldFilterInterface
     public function supportsField($field)
     {
         return in_array($field, $this->supportedFields);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
-    }
-
-    /**
-     * Prepare criteria condition with field, operator and value
-     *
-     * @param string|array $field    the backend field name
-     * @param string|array $operator the operator used to filter
-     * @param string|array $value    the value(s) to filter
-     *
-     * @return string
-     * @throws ProductQueryException
-     */
-    protected function prepareCriteriaCondition($field, $operator, $value)
-    {
-        $criteriaCondition = new CriteriaCondition($this->qb);
-
-        return $criteriaCondition->prepareCriteriaCondition($field, $operator, $value);
     }
 }
