+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Filter/StringFilter.php
@@ -2,12 +2,8 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Filter;
 
-use Doctrine\ORM\QueryBuilder;
-use Pim\Bundle\CatalogBundle\Doctrine\ORM\Condition\CriteriaCondition;
-use Pim\Bundle\CatalogBundle\Doctrine\ORM\Join\ValueJoin;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\Operators;
 use Pim\Bundle\CatalogBundle\Doctrine\Query\AttributeFilterInterface;
-use Pim\Bundle\CatalogBundle\Exception\ProductQueryException;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 
 /**
@@ -17,19 +13,13 @@ use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class StringFilter implements AttributeFilterInterface
+class StringFilter extends AbstractFilter implements AttributeFilterInterface
 {
-    /** @var QueryBuilder */
-    protected $qb;
-
     /** @var array */
     protected $supportedAttributes;
 
     /** @var array */
     protected $supportedFields;
-
-    /** @var array */
-    protected $supportedOperators;
 
     /**
      * Instanciate the base filter
@@ -46,18 +36,6 @@ class StringFilter implements AttributeFilterInterface
         $this->supportedAttributes = $supportedAttributes;
         $this->supportedFields     = $supportedFields;
         $this->supportedOperators  = $supportedOperators;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setQueryBuilder($queryBuilder)
-    {
-        if (!(($queryBuilder instanceof QueryBuilder))) {
-            throw new \InvalidArgumentException('Query builder should be an instance of Doctrine\ORM\QueryBuilder');
-        }
-
-        $this->qb = $queryBuilder;
     }
 
     /**
@@ -108,22 +86,6 @@ class StringFilter implements AttributeFilterInterface
     }
 
     /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
-    }
-
-    /**
      * Prepare conditions of the filter
      * @param string|array $operator
      * @param string|array $value
@@ -154,41 +116,5 @@ class StringFilter implements AttributeFilterInterface
         }
 
         return $this->prepareCriteriaCondition($backendField, $operator, $value);
-    }
-
-    /**
-     * Prepare criteria condition with field, operator and value
-     *
-     * @param string|array $field    the backend field name
-     * @param string|array $operator the operator used to filter
-     * @param string|array $value    the value(s) to filter
-     *
-     * @return string
-     * @throws ProductQueryException
-     */
-    protected function prepareCriteriaCondition($field, $operator, $value)
-    {
-        $criteriaCondition = new CriteriaCondition($this->qb);
-
-        return $criteriaCondition->prepareCriteriaCondition($field, $operator, $value);
-    }
-
-    /**
-     * Prepare join to attribute condition with current locale and scope criterias
-     *
-     * @param AttributeInterface $attribute the attribute
-     * @param string             $joinAlias the value join alias
-     * @param string             $locale    the locale
-     * @param string             $scope     the scope
-     *
-     * @throws ProductQueryException
-     *
-     * @return string
-     */
-    protected function prepareAttributeJoinCondition(AttributeInterface $attribute, $joinAlias, $locale = null, $scope = null)
-    {
-        $joinHelper = new ValueJoin($this->qb);
-
-        return $joinHelper->prepareCondition($attribute, $joinAlias, $locale, $scope);
     }
 }
